Refactor synthetic scopes
Previously, synthetic scopes live outside of scopes hierarchy.
This is an attempt to bring them into the hierarchy and simplify their
usage and extension.

TL;DR:
1) New class SyntheticResolutionScope
2) Old synthetic scopes are now called "providers"
3) New metadata for providers

Before the change, the was no way to get one single synthetic scope
which provide information about several synthetic entities, for instance,
properties and SAM contructors. The user code used to retrieve multiple
synthetic scope in order to get the info.

Now, the user code just needs to provide metadata alongside with scope
to cover to get all info it requires in one step.

For example, it the user code needs synthetic scope for properties and
SAM constructor, the code to retrieve the scope looks like:

syntheticScopes.provideSyntheticScope(
    type.memberScope,
    SyntheticScopeMetadata(
       needProperties = true,
       needConstructors = true
    )
).getContributedDescriptors()

Of course, SyntheticScopes' extension functions are still accessible
and they provide the same data as they used to, so, there is no change
in existing user code.

In order to create new Provider, one needs to inherit from
SyntheticResolutionScope and override the methods to provide new
entities in the scope; and SyntheticScopeProvider with method
provideSyntheticScope, where he looks at the metadata and decides
whether he needs to provide new scope or leave it unchanged.